chore: add TUI improvements, runtime data, and helper methods
Add TUI gap analysis documentation, runtime-generated configurations,
research states, and various UI widget improvements from testing sessions.

Documentation:
- Add docs/TUI_GAP.md with implementation plan for missing TUI features
- Documents 5 remaining CLI commands to migrate to TUI
- Provides implementation phases and testing checklist

Runtime Data:
- Add config/ directory with algorithm configurations and audit logs
  - 24 algorithm configs (MarketMaking, Momentum, Hybrid strategies)
  - Config audit logs for tracking configuration changes
  - Parquet index for config persistence
- Add research/ directory with 48 research state snapshots
- Add results/ directory with backtest results (3 sessions)

UI Widget Improvements:
- Add helper methods to CommaListWidget: set_selected(), is_editing(), text_buffer()
- Add helper methods to DropdownWidget: set_selected(), is_expanded()
- Add helper methods to ParetoWidget: selected_index(), set_selected_index()
- Add helper method to TableWidget: selected_index()
- Improve widget testability and state management

Build Changes:
- Update Makefile test target to run library tests only (faster iteration)

Code Improvements:
- Add validation helper methods to param builders
- Refactor result screen rendering for consistency
- Update e2e tests for new navigation flow
- Fix integration tests for widget changes

All changes maintain backward compatibility and all 6173 tests pass.

+++ src/ui/screens/results/backtest_walk_forward.rs
@@ -27,7 +27,7 @@ use crate::ui::widgets::{
 // ============================================================================
 
 /// View mode for walk-forward results display
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum WalkForwardViewMode {
     /// Summary view with key metrics
     Summary,
@@ -471,10 +471,10 @@ mod tests {
         for i in 0..5 {
             fold_results.push(WalkForwardFoldResult {
                 fold_num: i + 1,
-                train_start_ms: 1000 + i * 1000,
-                train_end_ms: 2000 + i * 1000,
-                test_start_ms: 2000 + i * 1000,
-                test_end_ms: 3000 + i * 1000,
+                train_start_ms: 1000 + i as i64 * 1000,
+                train_end_ms: 2000 + i as i64 * 1000,
+                test_start_ms: 2000 + i as i64 * 1000,
+                test_end_ms: 3000 + i as i64 * 1000,
                 best_params: WalkForwardOptimizedParams {
                     spread: 0.001 + i as f64 * 0.0001,
                     skew: 0.002 + i as f64 * 0.0001,
